+++ tests/test_capture.py
@@ -33,8 +33,6 @@ def test_create_poller(motion_controller):
         assert pytest.approx(tared_ts // 1 * 0.2) == value
 
 
-@pytest.mark.soem
-@pytest.mark.eoe
 def test_create_monitoring_no_trigger(motion_controller,
                                       disable_monitoring_disturbance):
     registers = [{
@@ -44,39 +42,38 @@ def test_create_monitoring_no_trigger(motion_controller,
     mc, alias = motion_controller
     mc.motion.set_operation_mode(OperationMode.VELOCITY, alias)
     max_frequency = mc.configuration.get_position_and_velocity_loop_rate(alias)
-    divider = 20
-    samples = 4000
+    divider = 40
+    samples = 2000
+    quarter_num_samples = samples // 4
     freq = max_frequency/divider
     total_time = samples/freq
-    monitoring = mc.capture.create_monitoring(registers, divider,
-                                              total_time, servo=alias)
+    quarter_total_time = total_time/4
     mc.motion.move_to_position(0, alias)
     mc.motion.motor_enable(alias)
+    monitoring = mc.capture.create_monitoring(registers, divider,
+                                              total_time, servo=alias)
     mc.capture.enable_monitoring_disturbance(servo=alias)
     init = time.time()
-    quarter_total_time = total_time/4
     for i in range(1, 4):
         while init + i*quarter_total_time > time.time():
             pass
-        mc.motion.move_to_position(1000*i, alias)
+        mc.motion.move_to_position(quarter_num_samples*i, alias)
     data = monitoring.read_monitoring_data()
     assert samples == len(data[0])
     for index, value in enumerate(data[0]):
-        subindex = index % 1000
-        theo_value = index//1000 * 1000
+        subindex = index % quarter_num_samples
+        theo_value = index//quarter_num_samples * quarter_num_samples
         if subindex > 100:  # Ignore first 100 samples for each value change
             assert value == theo_value
 
 
-@pytest.mark.soem
-@pytest.mark.eoe
 @pytest.mark.parametrize("trigger_mode, trigger_config, values_list", [
     (MonitoringSoCType.TRIGGER_EVENT_EDGE,
      MonitoringSoCConfig.TRIGGER_CONFIG_RISING,
-     [0, 1000, 2000, 3000]),
+     [0, 0.25, 0.5, 0.75]),
     (MonitoringSoCType.TRIGGER_EVENT_EDGE,
      MonitoringSoCConfig.TRIGGER_CONFIG_FALLING,
-     [3000, 2000, 1000, 0]),
+     [0.75, 0.5, 0.25, 0]),
 ])
 def test_create_monitoring_edge_trigger(motion_controller, trigger_mode, trigger_config,
                                         values_list, disable_monitoring_disturbance):
@@ -87,8 +84,10 @@ def test_create_monitoring_edge_trigger(motion_controller, trigger_mode, trigger
     mc, alias = motion_controller
     mc.motion.set_operation_mode(OperationMode.VELOCITY, alias)
     max_frequency = mc.configuration.get_position_and_velocity_loop_rate(alias)
-    divider = 20
-    samples = 4000
+    divider = 40
+    samples = 2000
+    quarter_num_samples = samples // 4
+    trigger_value = ((values_list[1] + values_list[2]) / 2) * samples
     freq = max_frequency/divider
     total_time = samples/freq
     monitoring = mc.capture.create_monitoring(
@@ -97,10 +96,10 @@ def test_create_monitoring_edge_trigger(motion_controller, trigger_mode, trigger
         trigger_mode=trigger_mode,
         trigger_config=trigger_config,
         trigger_signal=register,
-        trigger_value=1500,
+        trigger_value=trigger_value,
         servo=alias
     )
-    mc.motion.move_to_position(values_list[0], alias)
+    mc.motion.move_to_position(int(values_list[0] * samples), alias)
     mc.motion.motor_enable(alias)
     mc.capture.enable_monitoring_disturbance(servo=alias)
     time.sleep(1)
@@ -109,33 +108,29 @@ def test_create_monitoring_edge_trigger(motion_controller, trigger_mode, trigger
     for i in range(1, 4):
         while init + i*quarter_total_time > time.time():
             pass
-        mc.motion.move_to_position(values_list[i], alias)
+        mc.motion.move_to_position(int(values_list[i] * samples), alias)
     data = monitoring.read_monitoring_data()
     assert samples == len(data[0])
-    assert data[0][2000] == values_list[2]
-    assert data[0][1999] == values_list[1]
     for index, value in enumerate(data[0]):
-        subindex = index % 1000
-        theo_value = values_list[index//1000]
-        if 10 < subindex < 990:  # Ignore some samples near of value changes
+        subindex = index % quarter_num_samples
+        theo_value = int(values_list[index//quarter_num_samples] * samples)
+        if 10 < subindex < quarter_num_samples - 10:  # Ignore some samples near of value changes
             assert value == theo_value
 
 
-@pytest.mark.soem
-@pytest.mark.eoe
 def test_create_disturbance(motion_controller,
                             disable_monitoring_disturbance):
     mc, alias = motion_controller
     target_register = "CL_POS_SET_POINT_VALUE"
     max_frequency = mc.configuration.get_position_and_velocity_loop_rate(alias)
-    divider = 20
-    samples = 4000
+    divider = 40
+    samples = 2000
     freq = max_frequency/divider
     period = 1/freq
     data = []
-    data_subrange = 1000
+    data_subrange = samples // 4
     for i in range(samples//data_subrange):
-        data += [i * data_subrange] * 1000
+        data += [i * data_subrange] * data_subrange
     dist = mc.capture.create_disturbance(target_register, data, divider, servo=alias)
     init_time = time.time()
     mc.capture.enable_monitoring_disturbance(servo=alias)
@@ -143,13 +138,11 @@ def test_create_disturbance(motion_controller,
         time_now = time.time() - init_time
         current_value = mc.communication.get_register(target_register, alias)
         sample_num = int((time_now//period) % samples)
-        if sample_num % data_subrange < 10:
+        if sample_num % data_subrange < 20:
             continue
         assert current_value == data[sample_num]
 
 
-@pytest.mark.soem
-@pytest.mark.eoe
 @pytest.mark.smoke
 def test_mcb_synchronization(mocker, motion_controller):
     mc, alias = motion_controller
@@ -162,8 +155,6 @@ def test_mcb_synchronization(mocker, motion_controller):
     disable_mon.assert_called_once_with(servo=alias)
 
 
-@pytest.mark.soem
-@pytest.mark.eoe
 @pytest.mark.smoke
 def test_mcb_synchronization_fail(motion_controller):
     mc, alias = motion_controller